Moved ID generators for Poland to a country-specific package 'pol'
The convention is to use ISO3 country code as the package name.
This should make it easier to discover country-specific
generators if more are added in the future.
