+++ instancio-core/src/main/java/org/instancio/generators/IdGenerators.java
@@ -19,14 +19,9 @@ package org.instancio.generators;
 import org.instancio.generator.GeneratorContext;
 import org.instancio.generator.specs.EanGeneratorSpec;
 import org.instancio.generator.specs.IsbnGeneratorSpec;
-import org.instancio.generator.specs.NipGeneratorSpec;
-import org.instancio.generator.specs.PeselGeneratorSpec;
-import org.instancio.generator.specs.RegonGeneratorSpec;
+import org.instancio.generators.pol.PolIdGenerators;
 import org.instancio.internal.generator.domain.id.EanGenerator;
 import org.instancio.internal.generator.domain.id.IsbnGenerator;
-import org.instancio.internal.generator.domain.id.NipGenerator;
-import org.instancio.internal.generator.domain.id.PeselGenerator;
-import org.instancio.internal.generator.domain.id.RegonGenerator;
 
 /**
  * Contains built-in generators for various types of identifiers.
@@ -62,32 +57,12 @@ public class IdGenerators {
     }
 
     /**
-     * Generates Polish VAT Identification Number (NIP).
+     * Provides access to identifier generators for Poland.
      *
-     * @return API builder reference
+     * @return built-in id generators for Poland
      * @since 3.1.0
      */
-    public NipGeneratorSpec nip() {
-        return new NipGenerator(context);
-    }
-
-    /**
-     * Generates Polish National Identification Number (PESEL).
-     *
-     * @return API builder reference
-     * @since 3.1.0
-     */
-    public PeselGeneratorSpec pesel() {
-        return new PeselGenerator(context);
-    }
-
-    /**
-     * Generates Polish Taxpayer Identification Number (REGON).
-     *
-     * @return API builder reference
-     * @since 3.1.0
-     */
-    public RegonGeneratorSpec regon() {
-        return new RegonGenerator(context);
+    public PolIdGenerators pol() {
+        return new PolIdGenerators(context);
     }
 }
