+++ instancio-core/src/main/java/org/instancio/internal/beanvalidation/HibernateBeanValidationProcessor.java
@@ -35,9 +35,9 @@ import org.instancio.internal.generator.checksum.Mod11Generator;
 import org.instancio.internal.generator.domain.finance.CreditCardNumberGenerator;
 import org.instancio.internal.generator.domain.id.EanGenerator;
 import org.instancio.internal.generator.domain.id.IsbnGenerator;
-import org.instancio.internal.generator.domain.id.NipGenerator;
-import org.instancio.internal.generator.domain.id.PeselGenerator;
-import org.instancio.internal.generator.domain.id.RegonGenerator;
+import org.instancio.internal.generator.domain.id.pol.NipGenerator;
+import org.instancio.internal.generator.domain.id.pol.PeselGenerator;
+import org.instancio.internal.generator.domain.id.pol.RegonGenerator;
 import org.instancio.internal.generator.domain.internet.EmailGenerator;
 import org.instancio.internal.generator.net.URLGenerator;
 import org.instancio.internal.generator.util.UUIDGenerator;
