+++ instancio-core/src/main/java/org/instancio/internal/generator/domain/id/pol/WeightsModCheckGenerator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.instancio.internal.generator.domain.id;
+package org.instancio.internal.generator.domain.id.pol;
 
 import org.instancio.generator.GeneratorContext;
 import org.instancio.internal.generator.checksum.BaseModCheckGenerator;
