+++ instancio-tests/feature-tests/src/test/java/org/instancio/test/features/generator/id/pol/NipGeneratorTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.instancio.test.features.generator.id;
+package org.instancio.test.features.generator.id.pol;
 
 import org.instancio.GeneratorSpecProvider;
 import org.instancio.Instancio;
@@ -31,46 +31,28 @@ import static org.instancio.Select.root;
 
 @FeatureTag(Feature.GENERATOR)
 @ExtendWith(InstancioExtension.class)
-class RegonGeneratorTest {
+class NipGeneratorTest {
 
-    private static String createRegon(final GeneratorSpecProvider<String> spec) {
+    private static String create(final GeneratorSpecProvider<String> spec) {
         return Instancio.of(String.class)
                 .generate(root(), spec)
                 .create();
     }
 
     @Test
-    void defaultRegonIsType9() {
-        final String result = createRegon(gen -> gen.id().regon());
+    void nip() {
+        final String result = create(gen -> gen.id().pol().nip());
 
-        assertRegon(result, 9);
-    }
-
-    @Test
-    void regon9() {
-        final String result = createRegon(gen -> gen.id().regon().type9());
-
-        assertRegon(result, 9);
-    }
-
-    @Test
-    void regon14() {
-        final String result = createRegon(gen -> gen.id().regon().type14());
-
-        assertRegon(result, 14);
+        assertThat(result)
+                .hasSize(10)
+                .containsOnlyDigits();
     }
 
     @Test
     void nullable() {
-        final Stream<String> result = Stream.generate(() -> createRegon(gen -> gen.id().regon().nullable()))
+        final Stream<String> result = Stream.generate(() -> create(gen -> gen.id().pol().nip().nullable()))
                 .limit(Constants.SAMPLE_SIZE_DDD);
 
         assertThat(result).containsNull();
-    }
-
-    private static void assertRegon(final String result, final int expectedSize) {
-        assertThat(result)
-                .hasSize(expectedSize)
-                .containsOnlyDigits();
     }
 }
